move response checkers to utils library
